Bugfix: dtype

+++ src/rlenv/wrappers/paddings.py
@@ -21,7 +21,7 @@ class PadExtras(RLEnvWrapper):
         return self._add_extras(super().reset())
     
     def _add_extras(self, obs: Observation) -> Observation:
-        obs.extras = np.concatenate([obs.extras, np.zeros((obs.n_agents, self.n))], axis=-1)
+        obs.extras = np.concatenate([obs.extras, np.zeros((obs.n_agents, self.n), dtype=np.float32)], axis=-1)
         return obs
     
     def kwargs(self) -> dict[str,]:
@@ -48,7 +48,7 @@ class PadObservations(RLEnvWrapper):
         return self._add_obs(super().reset())
     
     def _add_obs(self, obs: Observation) -> Observation:
-        obs.data = np.concatenate([obs.data, np.zeros((obs.n_agents, self.n))], axis=-1)
+        obs.data = np.concatenate([obs.data, np.zeros((obs.n_agents, self.n), dtype=np.float32)], axis=-1)
         return obs
 
     def kwargs(self) -> dict[str,]:
